# src/main/resources/openapi.yaml
...
openapi: "3.0.4"
...
info:
  title: Beveage Service
  version: "1.0"
  description: | # | for multi line description
    Summer is coming and we are all looking forward to having a barbecue with friends.
    For a nice barbecue, we also need some refreshing drinks. Therefore, we want to start a
    new business where we sell beverages and deliver them to people.
  contact:
    name: Group 9
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:9999/v1
    description: Development Server
# resources (one of the REST principles)
...
components:
  schemas:
    orderDTO:
      type: object
      required:
        - orderId
      properties:
        orderId:
          type: number
        positions:
          type: array
          items:
            $ref: "#/components/schemas/orderItemDTO"
        price:
          type: number
        OrderStatus:
          type: string
          enum: [SUBMITTED, PROCESSED]
      xml:
        name: order
    orderItemDTO:
      type: object
      required:
        - number
        - beverage
        - quantity
      properties:
        number:
          type: number
        beverage:
          $ref: "#/components/schemas/beverageDTO"
        quantity:
          type: number
      xml:
        name: orderItem
    beverageDTO:
      type: object
      required:
        - Bottle
      properties:
        id:
          type: number
        price:
          type: number
        inStock:
          type: boolean
        Bottle:
          $ref: "#/components/schemas/beverageDTO"
        noOfBottles:
          type: number
      xml:
        name: beverage
...
    errorMessage:
      type: object
      required:
        - errorType
      properties:
        errorType:
          description: error category, should be extended
          enum:
            - INVALID_PARAMETER
        message:
          type: string
          description: A detailed error description
    pagination:
      type: object
      properties:
        page:
          type: number
        noOfPages:
          type: number
        first:
          type: string
        previous:
          type: string
        next:
          type: string
        last:
          type: string
...
    paginatedOrders:
      type: object
      properties:
        pagination:
          $ref: "#/components/schemas/pagination"
        beverages:
          type: array
          items:
            $ref: "#/components/schemas/orderDTO"
        href:
          type: string
...
paths:
...
  /order:
    description: A catalog of all orders
    # Operations/methods via standard HTTP verbs
    get:
      tags:
        - shared
      summary: Get order on a specific page
      # request parameters
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: pageLimit
          in: query
          schema:
            type: number
            default: 10
        - name: search
          in: query
          description: simple string contains search [not case sensitive]
          schema:
            type: string
        - name: filter
          in: query
          description: The filter/Sort function you want to apply
          schema:
            type: string
            enum: [MaxPrice, MinPrice, OrderProcessed, OrderSubmitted]
      # possible responses and their parameters
      responses:
        200:
          description: orders on the page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/paginatedOrders"
        400:
          description: A bad request initiated by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
    post:
      tags:
        - customer
      summary: Create an order
      responses:
        201:
          description: order created
        400:
          description: Message body was empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /order/{orderId}:
    description: delete an order
    parameters:
      - name: orderId
        # path parameters must be always mandatory!!
        required: true
        in: path
        schema:
          type: number
    get:
      tags:
        - shared
      summary: Get a single order by id
      # possible responses and their parameters
      responses:
        200:
          description: get a specific order description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orderDTO"
        400:
          description: A bad request initiated by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
    put:
      tags:
        - employee
      summary: set order to processed
      responses:
        200:
          description: Updating an order
          content:
            application/json:
              # response schema
              schema:
                $ref: "#/components/schemas/orderDTO"
        400:
          description: Message body was empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        404:
          description: Not found
    delete:
      tags:
        - employee
      summary: Delete order
      responses:
        200:
          description: order deleted
        404:
          description: order not found
...